http://arcemu.org/forums/index.php?showtopic=1543&hl= -> - implemented CMSG_PET_SPELL_AUTOCAST and CMSG_PET_CANCEL_AURA - proper SMSG for pet spell failure - dbc based HP for revived pet - pet do not loose happiness in BGs - you cant feed your dead pet

+++ src/arcemu-world/Player.h
@@ -1069,7 +1069,7 @@ public:
 	void						SpawnPet(uint32 pet_number);
 	void						DespawnPet();
 
-	ARCEMU_INLINE uint8            GetPetCount(void) { return m_Pets.size(); }
+	ARCEMU_INLINE uint8         GetPetCount(void) { return (uint8)m_Pets.size(); }
 	ARCEMU_INLINE void			SetStableSlotCount(uint8 count) { m_StableSlotCount = count; }
 	ARCEMU_INLINE uint8			GetStableSlotCount(void) { return m_StableSlotCount; }
 	uint32						GetUnstabledPetNumber(void)
